shared preferences basic functionality commit
Uses an intent with extra data to start a new activity to save the open
the shared preferences editor and save the player game stats then ends
the activity and returns to the main activity which picked up the data
contained in the shared preferences object. more functionalities to be
added later as pending projects exists.